if there is an API key but no settings category, prepend it
If we can't find `[settings] in the config file but we can find an API key, we'll write the `[settings]` line to a new line of an auxiliary file and append `infile` to it, then copy the whole aux file to the `infile`. This way we'll get the `[settings]` category at the beginning of the file.